pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "barqhwmusig"
version = "0.1.0"
description = "Bitcoin Multisig POC with Hardware Wallet Integration"
readme = "README.md"
requires-python = ">=3.14"
license = {text = "MIT"}
authors = [
    {name = "BarqHWMuSig Team"},
]
dependencies = [
    "python-bitcoinlib>=0.11.0",
    "yubikey-manager>=4.0.0",
    "fido2>=1.0.0",
    "pyusb>=1.2.1",
    "btchip-python>=0.1.32",
    "cryptography>=39.0.0",
    "click>=8.1.3",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.1.0",
    "coverage>=7.2.5",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "ruff>=0.0.262",
    "bandit>=1.7.5",
    "pre-commit>=3.3.1",
    "pip-audit>=2.5.5",
]

[tool.black]
line-length = 88
target-version = ["py314"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py314"
select = ["E", "F", "B", "I", "N", "UP", "ANN", "S", "A"]
ignore = ["ANN101"]  # Missing type annotation for `self`

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=src --cov-report=term-missing" 